bug annotations crash
git-svn-id: https://svn.code.sf.net/p/osirix/code/osirix@2945 4e9e6e3d-4551-47eb-9dbe-a6127f9400a1

+++ DICOMExport.h
@@ -67,7 +67,7 @@
 - (void) setSeriesNumber: (long) no;
 - (void) setDefaultWWWL: (long) ww :(long) wl;
 - (void) setPixelSpacing: (float) x :(float) y;
-- (void) setSliceThickness: (float) t;
+- (void) setSliceThickness: (double) t;
 - (void) setOrientation: (float*) o;
 - (void) setPosition: (float*) p;
 - (void) setSlicePosition: (float) p;
